test: split provider mocks by driver like the import managers
mocks/google and mocks/microsoft mirror the drivers/ layout on the
business side: one handler module per API surface (token, identity,
mailbox, calendar events), composed by the provider's setup.
