+++ packages/twenty-server/test/integration/calendar/suites/calendar-stale-sync.integration-spec.ts
@@ -8,7 +8,7 @@ import { CalendarOngoingStaleCronJob } from 'src/modules/calendar/calendar-event
 
 import { connectMessagingAccount } from 'test/integration/messaging/utils/connect-messaging-account.util';
 import { queryCalendarChannel } from 'test/integration/messaging/utils/query-messaging.util';
-import { setupGoogleMock } from 'test/integration/mocks/setup-google-mock.util';
+import { setupGoogleMock } from 'test/integration/mocks/google/setup-google-mock.util';
 import { getCoreRepository } from 'test/integration/utils/get-core-repository.util';
 import { runSyncCron } from 'test/integration/utils/run-sync-cron.util';
