+++ packages/twenty-server/test/integration/calendar/suites/calendar-sync-failure-lifecycle.integration-spec.ts
@@ -16,7 +16,7 @@ import {
   queryConnectedAccount,
   updateCalendarChannel,
 } from 'test/integration/messaging/utils/query-messaging.util';
-import { setupGoogleMock } from 'test/integration/mocks/setup-google-mock.util';
+import { setupGoogleMock } from 'test/integration/mocks/google/setup-google-mock.util';
 import { getCoreRepository } from 'test/integration/utils/get-core-repository.util';
 import { runSyncCron } from 'test/integration/utils/run-sync-cron.util';
