+++ packages/twenty-server/test/integration/messaging/suites/messaging-incremental-sync.integration-spec.ts
@@ -1,9 +1,9 @@
 import { ConnectedAccountProvider } from 'twenty-shared/types';
 
 import { connectMessagingAccount } from 'test/integration/messaging/utils/connect-messaging-account.util';
-import { getGmailMessageSubject } from 'test/integration/mocks/gmail-message-subject.util';
-import { gmailMessage } from 'test/integration/mocks/gmail-message.util';
-import { setupGoogleMock } from 'test/integration/mocks/setup-google-mock.util';
+import { getGmailMessageSubject } from 'test/integration/mocks/google/gmail-message-subject.util';
+import { gmailMessage } from 'test/integration/mocks/google/gmail-message.util';
+import { setupGoogleMock } from 'test/integration/mocks/google/setup-google-mock.util';
 import { findImportedMessageSubjects } from 'test/integration/utils/find-imported-records.util';
 import { runMessageChannelSync } from 'test/integration/utils/run-message-channel-sync.util';
