+++ packages/twenty-server/test/integration/messaging/suites/messaging-token-refresh.integration-spec.ts
@@ -4,8 +4,8 @@ import { ConnectedAccountEntity } from 'src/engine/metadata-modules/connected-ac
 
 import { connectMessagingAccount } from 'test/integration/messaging/utils/connect-messaging-account.util';
 import { queryConnectedAccount } from 'test/integration/messaging/utils/query-messaging.util';
-import { gmailMessage } from 'test/integration/mocks/gmail-message.util';
-import { setupGoogleMock } from 'test/integration/mocks/setup-google-mock.util';
+import { gmailMessage } from 'test/integration/mocks/google/gmail-message.util';
+import { setupGoogleMock } from 'test/integration/mocks/google/setup-google-mock.util';
 import { getCoreRepository } from 'test/integration/utils/get-core-repository.util';
 import { runMessageChannelSync } from 'test/integration/utils/run-message-channel-sync.util';
