+++ packages/twenty-server/test/integration/mocks/google/google-calendar-events-url.constant.ts
@@ -1,0 +1,2 @@
+export const GOOGLE_CALENDAR_EVENTS_URL =
+  'https://www.googleapis.com/calendar/v3/calendars/primary/events';
